refactor: solve theme circular dependency issue
