+++ components/theme/interface/index.ts
@@ -1,5 +1,6 @@
-import type { ComponentTokenMap } from './components';
+import { CSSInterpolation } from '@ant-design/cssinjs';
 import type { AliasToken } from './alias';
+import type { ComponentTokenMap } from './components';
 
 export type OverrideToken = {
   [key in keyof ComponentTokenMap]: Partial<ComponentTokenMap[key]> & Partial<AliasToken>;
@@ -8,23 +9,30 @@ export type OverrideToken = {
 /** Final token which contains the components level override */
 export type GlobalToken = AliasToken & ComponentTokenMap;
 
-export { PresetColors } from './presetColors';
+export type { AliasToken } from './alias';
+export type { ComponentTokenMap } from './components';
 export type {
-  PresetColorType,
-  ColorPalettes,
-  LegacyColorPalettes,
-  PresetColorKey,
-} from './presetColors';
-export type { SeedToken } from './seeds';
-export type {
-  MapToken,
   ColorMapToken,
   ColorNeutralMapToken,
   CommonMapToken,
-  HeightMapToken,
-  SizeMapToken,
   FontMapToken,
+  HeightMapToken,
+  MapToken,
+  SizeMapToken,
   StyleMapToken,
 } from './maps';
-export type { AliasToken } from './alias';
-export type { ComponentTokenMap } from './components';
+export { PresetColors } from './presetColors';
+export type {
+  ColorPalettes,
+  LegacyColorPalettes,
+  PresetColorKey,
+  PresetColorType,
+} from './presetColors';
+export type { SeedToken } from './seeds';
+
+export type UseComponentStyleResult = [(node: React.ReactNode) => React.ReactElement, string];
+
+export type GenerateStyle<
+  ComponentToken extends object = AliasToken,
+  ReturnType = CSSInterpolation,
+> = (token: ComponentToken) => ReturnType;
