+++ webpack.config.js
@@ -3,6 +3,7 @@
 const getWebpackConfig = require('@ant-design/tools/lib/getWebpackConfig');
 const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 const { EsbuildPlugin } = require('esbuild-loader');
+const CircularDependencyPlugin = require('circular-dependency-plugin');
 const DuplicatePackageCheckerPlugin = require('duplicate-package-checker-webpack-plugin');
 
 function addLocales(webpackConfig) {
@@ -57,6 +58,12 @@ if (process.env.RUN_ENV === 'PRODUCTION') {
         }),
       );
     }
+    config.plugins.push(
+      new CircularDependencyPlugin({
+        // add errors to webpack instead of warnings
+        failOnError: true,
+      }),
+    );
   });
 }
